remove client logs, revert start script, add * endpoint
