+++ package.json
@@ -69,8 +69,7 @@
     "stateful-react-youtube": "^0.0.6"
   },
   "scripts": {
-    "start": "node server/server.js",
-    "start-dev": "node scripts/start.js",
+    "start": "node scripts/start.js",
     "build": "node scripts/build.js",
     "test": "node scripts/test.js --env=jsdom"
   },
