+++ public/index.html
@@ -30,6 +30,5 @@
       To create a production bundle, use `npm run build`.
     -->
     <script src="https://www.youtube.com/player_api"></script>
-    <script src="http://localhost:8080/socket.io/socket.io.js"></script>
   </body>
 </html>
