+++ scripts/start.js
@@ -32,16 +32,16 @@ var handleCompile;
 
 // You can safely remove this after ejecting.
 // We only use this block for testing of Create React App itself:
-var isSmokeTest = process.argv.some(arg => arg.indexOf('--smoke-test') > -1);
-if (isSmokeTest) {
-  handleCompile = function (err, stats) {
-    if (err || stats.hasErrors() || stats.hasWarnings()) {
-      process.exit(1);
-    } else {
-      process.exit(0);
-    }
-  };
-}
+// var isSmokeTest = process.argv.some(arg => arg.indexOf('--smoke-test') > -1);
+// if (isSmokeTest) {
+//   handleCompile = function (err, stats) {
+//     if (err || stats.hasErrors() || stats.hasWarnings()) {
+//       process.exit(1);
+//     } else {
+//       process.exit(0);
+//     }
+//   };
+// }
 
 function setupCompiler(host, port, protocol) {
   // "Compiler" is a low-level interface to Webpack.
